move all bulk operation functionality into models - remove bulk operation services
