+++ app/jobs/bulk_operation/revert_applications_to_pending_job.rb
@@ -4,7 +4,7 @@ class BulkOperation::RevertApplicationsToPendingJob < ApplicationJob
   def perform(bulk_operation_id:)
     bulk_operation = BulkOperation::RevertApplicationsToPending.find(bulk_operation_id)
     Rails.logger.info("Bulk Operation started - bulk_operation_id: #{bulk_operation_id}")
-    BulkOperation::BulkRevertApplicationsToPending.new(bulk_operation:).run!(dry_run: false)
+    bulk_operation.run!
     Rails.logger.info("Bulk Operation finished - bulk_operation_id: #{bulk_operation_id}")
   end
 end
