+++ app/jobs/bulk_operation/update_and_verify_trns_job.rb
@@ -4,7 +4,7 @@ class BulkOperation::UpdateAndVerifyTrnsJob < ApplicationJob
   def perform(bulk_operation_id:)
     bulk_operation = BulkOperation::UpdateAndVerifyTrns.find(bulk_operation_id)
     Rails.logger.info("Bulk Operation started - bulk_operation_id: #{bulk_operation_id}")
-    BulkOperation::BulkUpdateAndVerifyTrns.new(bulk_operation:).run!
+    bulk_operation.run!
     Rails.logger.info("Bulk Operation finished - bulk_operation_id: #{bulk_operation_id}")
   end
 end
