--- app/services/bulk_operation/bulk_reject_applications.rb
@@ -1,37 +1,0 @@
-# frozen_string_literal: true
-
-class BulkOperation::BulkRejectApplications
-  attr_reader :bulk_operation
-
-  def initialize(bulk_operation:)
-    @bulk_operation = bulk_operation
-  end
-
-  def run!
-    result = {}
-    ActiveRecord::Base.transaction do
-      result = bulk_operation.ids_to_update.index_with do |application_ecf_id|
-        process_csv_row(application_ecf_id)
-      end
-      bulk_operation.update!(result: result.to_json, finished_at: Time.zone.now)
-    end
-
-    result
-  end
-
-private
-
-  def process_csv_row(application_ecf_id)
-    application = Application.find_by(ecf_id: application_ecf_id)
-    reject_service = Applications::Reject.new(application:)
-    success = reject_service.reject
-    outcome(success, application, reject_service.errors)
-  end
-
-  def outcome(success, application, errors)
-    return "Not found" if application.nil?
-    return "Changed to rejected" if success
-
-    errors.full_messages.to_sentence
-  end
-end
