+++ config/routes.rb
@@ -292,9 +292,11 @@ Rails.application.routes.draw do
         resources :revert_applications_to_pending, controller: "revert_applications_to_pending", only: %i[index create show] do
           post "run", on: :member
         end
+
         resources :reject_applications, controller: "reject_applications", only: %i[index create show] do
           post "run", on: :member
         end
+
         resources :update_and_verify_trns, controller: "update_and_verify_trns", only: %i[index create show] do
           post "run", on: :member
         end
