+++ spec/features/admin/bulk_operations/update_and_verify_trns_spec.rb
@@ -38,71 +38,86 @@ RSpec.feature "update and verify TRNs", :rack_test_driver, type: :feature do
 
   before do
     create :cohort, :current
-    sign_in_as_admin
     create(:user)
     create(:user)
   end
 
-  scenario "updating and verifying TRNs" do
-    visit npq_separation_admin_path
-    click_link "Bulk operations"
-    click_link "Update and verify TRNs"
-
-    expect(page).to have_content "No files have been uploaded"
-
-    attach_file "file", trns_file.path
-    click_button "Upload file"
-
-    expect(page).to have_content "File #{filename} uploaded successfully"
-
-    click_link filename
-    within(".govuk-summary-list") do |summary_list|
-      expect(summary_list).to have_summary_item("Filename", filename)
-      expect(summary_list).to have_summary_item("Rows", "2")
-      expect(summary_list).to have_summary_item("Created by", "#{admin.full_name} (#{admin.email})")
+  context "when not logged in" do
+    scenario "bulk operations index page is inaccessible" do
+      visit npq_separation_admin_bulk_operations_path
+      expect(page).to have_current_path(sign_in_path)
     end
 
-    visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
-    perform_enqueued_jobs do
-      click_button "Update and verify TRNs"
+    scenario "update and verify TRNs page is inaccessible" do
+      visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
+      expect(page).to have_current_path(sign_in_path)
     end
-
-    click_link filename
-    within(".govuk-summary-list") do |summary_list|
-      expect(summary_list).to have_summary_item("Filename", filename)
-      expect(summary_list).to have_summary_item("Rows", "2")
-      expect(summary_list).to have_summary_item("Created by", "#{admin.full_name} (#{admin.email})")
-      expect(summary_list).to have_summary_item("Ran by", "#{admin.full_name} (#{admin.email})")
-    end
-
-    expect(page).to have_content "#{User.first.ecf_id}TRN updated and verified"
-    expect(page).to have_content "#{User.last.ecf_id}TRN updated and verified"
-    expect(page).not_to have_button("Update and verify TRNs")
-    expect(User.all.pluck(:trn)).to eq %w[1234567 2345678]
-    expect(User.all.pluck(:trn_verified)).to eq [true, true]
   end
 
-  scenario "file validation" do
-    visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
-    attach_file "file", empty_file.path
-    click_button "Upload"
-    expect(page).to have_content "is empty"
+  context "when logged in as admin" do
+    before { sign_in_as(admin) }
 
-    visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
-    attach_file "file", only_headers_trns_file.path
-    click_button "Upload"
-    expect(page).to have_content "is empty"
+    scenario "updating and verifying TRNs" do
+      visit npq_separation_admin_path
+      click_link "Bulk operations"
+      click_link "Update and verify TRNs"
 
-    attach_file "file", wrong_format_trns_file.path
-    click_button "Upload"
-    expect(page).to have_content "is wrong format"
+      expect(page).to have_content "No files have been uploaded"
 
-    attach_file "file", no_headers_trns_file.path
-    click_button "Upload"
-    expect(page).to have_content "is wrong format"
+      attach_file "file", trns_file.path
+      click_button "Upload file"
 
-    attach_file "file", trns_file.path
-    click_button "Upload"
-    expect(page).to have_button "Update and verify TRNs"
+      expect(page).to have_content "File #{filename} uploaded successfully"
+
+      click_link filename
+      within(".govuk-summary-list") do |summary_list|
+        expect(summary_list).to have_summary_item("Filename", filename)
+        expect(summary_list).to have_summary_item("Rows", "2")
+        expect(summary_list).to have_summary_item("Created by", "#{admin.full_name} (#{admin.email})")
+      end
+
+      visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
+      perform_enqueued_jobs do
+        click_button "Update and verify TRNs"
+      end
+
+      click_link filename
+      within(".govuk-summary-list") do |summary_list|
+        expect(summary_list).to have_summary_item("Filename", filename)
+        expect(summary_list).to have_summary_item("Rows", "2")
+        expect(summary_list).to have_summary_item("Created by", "#{admin.full_name} (#{admin.email})")
+        expect(summary_list).to have_summary_item("Ran by", "#{admin.full_name} (#{admin.email})")
+      end
+
+      expect(page).to have_content "#{User.first.ecf_id}TRN updated and verified"
+      expect(page).to have_content "#{User.last.ecf_id}TRN updated and verified"
+      expect(page).not_to have_button("Update and verify TRNs")
+      expect(User.all.pluck(:trn)).to eq %w[1234567 2345678]
+      expect(User.all.pluck(:trn_verified)).to eq [true, true]
+    end
+
+    scenario "file validation" do
+      visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
+      attach_file "file", empty_file.path
+      click_button "Upload"
+      expect(page).to have_content "is empty"
+
+      visit npq_separation_admin_bulk_operations_update_and_verify_trns_path
+      attach_file "file", only_headers_trns_file.path
+      click_button "Upload"
+      expect(page).to have_content "is empty"
+
+      attach_file "file", wrong_format_trns_file.path
+      click_button "Upload"
+      expect(page).to have_content "is wrong format"
+
+      attach_file "file", no_headers_trns_file.path
+      click_button "Upload"
+      expect(page).to have_content "is wrong format"
+
+      attach_file "file", trns_file.path
+      click_button "Upload"
+      expect(page).to have_button "Update and verify TRNs"
+    end
   end
 end
